updated short room code
